sw_services: xilplmi: Fixes for MISRA C violation Rule 10.3
MISRA C Rule 10.3 States: The value of an expression shall not be assigned to an
object with a narrower essential type or of a different essential type category

This patch fixes MISRA C Rule 10.3 violation by using the proper type variables
where required or typecasting if necessary


Acked-for-series: Vikram Sreenivasa Batchali <b.vikram@xilinx.com>

+++ lib/sw_services/xilplmi/src/xplmi_debug.c
@@ -135,14 +135,14 @@ void outbyte(char c)
 {
 #ifdef STDOUT_BASEADDRESS
 	if(((LpdInitialized) & UART_INITIALIZED) == UART_INITIALIZED) {
-		XUartPsv_SendByte(STDOUT_BASEADDRESS, c);
+		XUartPsv_SendByte(STDOUT_BASEADDRESS, (u8)c);
 	}
 #endif
 
 	if (DebugLog.LogBuffer.CurrentAddr >=
 			(DebugLog.LogBuffer.StartAddr + DebugLog.LogBuffer.Len)) {
 		DebugLog.LogBuffer.CurrentAddr = DebugLog.LogBuffer.StartAddr;
-		DebugLog.LogBuffer.IsBufferFull = TRUE;
+		DebugLog.LogBuffer.IsBufferFull = (u8)TRUE;
 	}
 
 	XPlmi_OutByte64(DebugLog.LogBuffer.CurrentAddr, c);
