+++ lib/sw_services/xilplmi/src/xplmi_err.c
@@ -78,7 +78,7 @@ void XPlmi_ErrMgr(int ErrStatus)
 
 	/* Print the PLM error */
 	XPlmi_Printf(DEBUG_GENERAL, "PLM Error Status: 0x%08lx\n\r", ErrStatus);
-	XPlmi_Out32(PMC_GLOBAL_PMC_FW_ERR, ErrStatus);
+	XPlmi_Out32(PMC_GLOBAL_PMC_FW_ERR, (u32)ErrStatus);
 
 	/*
 	 * Fallback if boot PDI is not done
@@ -411,7 +411,7 @@ static void XPlmi_ErrPSMIntrHandler(u32 ErrorNodeId, u32 ErrorMask)
 	Err2Status = XPlmi_In32(PSM_GLOBAL_REG_PSM_ERR2_STATUS);
 
 	if (Err1Status != 0U) {
-		for (Index = XPLMI_NODEIDX_ERROR_PS_SW_CR;
+		for (Index = (u32)XPLMI_NODEIDX_ERROR_PS_SW_CR;
 				Index < XPLMI_NODEIDX_ERROR_PSMERR1_MAX; Index++) {
 			if (((Err1Status & ((u32)1U << (Index - (u32)XPLMI_NODEIDX_ERROR_PS_SW_CR))) != (u32)FALSE)
 				&& (ErrorTable[Index].Action != XPLMI_EM_ACTION_NONE)
@@ -422,7 +422,7 @@ static void XPlmi_ErrPSMIntrHandler(u32 ErrorNodeId, u32 ErrorMask)
 		}
 	}
 	if (Err2Status != 0U) {
-		for (Index = XPLMI_NODEIDX_ERROR_LPD_SWDT;
+		for (Index = (u32)XPLMI_NODEIDX_ERROR_LPD_SWDT;
 				Index < XPLMI_NODEIDX_ERROR_PSMERR2_MAX; Index++) {
 			if (((Err2Status & ((u32)1U << (Index - (u32)XPLMI_NODEIDX_ERROR_LPD_SWDT))) != (u32)FALSE)
 				&& (ErrorTable[Index].Action != XPLMI_EM_ACTION_NONE)
@@ -510,7 +510,7 @@ void XPlmi_ErrIntrHandler(void *CallbackRef)
 	 */
 
 	if (Err1Status != 0U) {
-		for (Index = XPLMI_NODEIDX_ERROR_BOOT_CR;
+		for (Index = (u32)XPLMI_NODEIDX_ERROR_BOOT_CR;
 				Index < XPLMI_NODEIDX_ERROR_PMCERR1_MAX; Index++) {
 			if (((Err1Status & ((u32)1U << Index)) != (u32)FALSE) &&
 				(ErrorTable[Index].Handler != NULL) &&
@@ -529,7 +529,7 @@ void XPlmi_ErrIntrHandler(void *CallbackRef)
 	}
 
 	if (Err2Status != 0U) {
-		for (Index = XPLMI_NODEIDX_ERROR_PMCAPB;
+		for (Index = (u32)XPLMI_NODEIDX_ERROR_PMCAPB;
 				Index < XPLMI_NODEIDX_ERROR_PMCERR2_MAX; Index++) {
 			if (((Err2Status &
 				((u32)1U << (Index - XPLMI_NODEIDX_ERROR_PMCERR1_MAX))) != (u32)FALSE)
@@ -986,7 +986,7 @@ void XPlmi_EmInit(s32 (* SystemShutdown)(u32 SubsystemId,
 	PmSystemShutdown = SystemShutdown;
 
 	/* Set the default actions as defined in the Error table */
-	for (Index = XPLMI_NODEIDX_ERROR_BOOT_CR;
+	for (Index = (u32)XPLMI_NODEIDX_ERROR_BOOT_CR;
 		Index < XPLMI_NODEIDX_ERROR_PMCERR1_MAX; Index++) {
 		if ((ErrorTable[Index].Action != XPLMI_EM_ACTION_INVALID) &&
 			(XPlmi_EmSetAction(XPLMI_EVENT_ERROR_PMC_ERR1, Index,
@@ -998,7 +998,7 @@ void XPlmi_EmInit(s32 (* SystemShutdown)(u32 SubsystemId,
 		}
 	}
 
-	for (Index = XPLMI_NODEIDX_ERROR_PMCAPB;
+	for (Index = (u32)XPLMI_NODEIDX_ERROR_PMCAPB;
 		Index < XPLMI_NODEIDX_ERROR_PMCERR2_MAX; Index++) {
 		if ((ErrorTable[Index].Action != XPLMI_EM_ACTION_INVALID) &&
 			(XPlmi_EmSetAction(XPLMI_EVENT_ERROR_PMC_ERR2, Index,
@@ -1036,7 +1036,7 @@ int XPlmi_PsEmInit(void)
 	XPlmi_Out32(PSM_GLOBAL_REG_PSM_ERR2_STATUS, MASK32_ALL_HIGH);
 
 	/* Set the default actions as defined in the Error table */
-	for (Index = XPLMI_NODEIDX_ERROR_PS_SW_CR;
+	for (Index = (u32)XPLMI_NODEIDX_ERROR_PS_SW_CR;
 		Index < XPLMI_NODEIDX_ERROR_PSMERR1_MAX; Index++) {
 		if ((ErrorTable[Index].Action != XPLMI_EM_ACTION_INVALID) &&
 			(XPlmi_EmSetAction(XPLMI_EVENT_ERROR_PSM_ERR1, Index,
@@ -1048,7 +1048,7 @@ int XPlmi_PsEmInit(void)
 		}
 	}
 
-	for (Index = XPLMI_NODEIDX_ERROR_LPD_SWDT;
+	for (Index = (u32)XPLMI_NODEIDX_ERROR_LPD_SWDT;
 		Index < XPLMI_NODEIDX_ERROR_PSMERR2_MAX; Index++) {
 		if ((ErrorTable[Index].Action != XPLMI_EM_ACTION_INVALID) &&
 			(XPlmi_EmSetAction(XPLMI_EVENT_ERROR_PSM_ERR2, Index,
